+++ ofspectrum/models/__init__.py
@@ -2,13 +2,14 @@
 OfSpectrum SDK Data Models
 """
 
-from .token import Token, TokenCreateParams, TokenUpdateParams
+from .token import AiAuthTag, Token, TokenCreateParams, TokenUpdateParams
 from .notebook import Notebook, NotebookMedia, NotebookCreateParams
 from .audio import EncodeResult, DecodeResult, StreamingEncodeResult
 from .quota import Quota, QuotaList
 
 __all__ = [
     "Token",
+    "AiAuthTag",
     "TokenCreateParams",
     "TokenUpdateParams",
     "Notebook",